ms2/pyproject.toml
[project]
name = "neurostream-ms2"
version = "0.1.0"
description = "AI perception service for NeuroStream"
requires-python = ">=3.11"
dependencies = [
    "boto3",
    "celery[redis]",
    "fastapi",
    "google-generativeai",
    "httpx",
    "openai-whisper",
    "pydantic",
    "python-dotenv",
    "uvicorn",
]

[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"

[tool.uv]
package = false

